requirements.txt
pymongo
pandas
flask_restplus
pyyaml
ipaddress
tldextract
scikit-learn